refactor: Refactor stucture
구조 개선을 위해 코드 리펙토링, play, appControl로 구분하여 게임 구성 객체들과 게임을 진행하는 요소들로 재구성함
